tests/wpt/css-tests/css-ui-3_dev/html/caret-color-020.htm

<!DOCTYPE html>
<html><head><meta charset="utf-8">
<title>CSS Basic User Interface Test: caret-color currentcolor test animation</title>
<link href="mailto:rego@igalia.com" rel="author" title="Manuel Rego Casasnovas">
<link href="http://www.w3.org/TR/css3-ui/#caret-color" rel="help">
<link href="https://www.w3.org/TR/web-animations-1/#dom-animatable-animate" rel="help">
<link href="https://www.w3.org/TR/css3-color/#color0" rel="help">
<meta content="Test checks that 'currentcolor' value for caret-color property is interpolable." name="assert">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
  textarea {
    color: magenta;
  }
</style>
</head><body>
  <textarea id="textarea"></textarea>
  <div id="log"></div>

  <script>
    test(
      function(){
        var textarea = document.getElementById("textarea");
        var keyframes = [
          { caretColor: 'currentcolor' },
          { caretColor: 'lime' }
        ];
        var options = {
          duration: 10,
          fill: "forwards"
        };

        var player = textarea.animate(keyframes, options);
        player.pause();
        player.currentTime = 5;
        assert_equals(getComputedStyle(textarea).caretColor, 'rgb(128, 128, 128)');
      }, "caret-color: currentcolor is interpolable");
</script>

</body></html>
